ZZIRIT-FLASK/requirements.txt
# Flask 핵심
Flask>=2.2.5
flask-cors>=3.0.10
gunicorn>=21.2.0

# 데이터 처리
pandas>=2.0.3
numpy>=1.25.2
openpyxl>=3.1.2

# 머신러닝
scikit-learn>=1.2.2
torch>=2.0.0
torchvision>=0.15.0
joblib>=1.3.0

# NLP/임베딩
sentence-transformers>=2.2.2

# API 및 네트워크
requests>=2.31.0
aiohttp>=3.8.0

# AI API
google-generativeai>=0.5.0

# 데이터베이스
pymysql>=1.1.0
mysql-connector-python>=8.1.0

# 클라우드/이메일
boto3

# 환경 관리
python-dotenv>=1.0.0

# 이미지 처리
Pillow>=10.0.0

# 인코딩/유틸리티
chardet>=5.1.0